+++ Hayden/Models/Yotsuba.cs
@@ -96,10 +96,10 @@ namespace Hayden.Models
 		public ushort? ImageHeight { get; set; }
 
 		[JsonProperty("tn_w")]
-		public byte? ThumbnailWidth { get; set; }
+		public ushort? ThumbnailWidth { get; set; }
 
 		[JsonProperty("tn_h")]
-		public byte? ThumbnailHeight { get; set; }
+		public ushort? ThumbnailHeight { get; set; }
 
 		[JsonConverter(typeof(BoolIntConverter))]
 		[JsonProperty("filedeleted")]
